requirements.txt
MouseInfo==0.1.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python3-xlib==0.15
pytweening==1.0.4
